Create CommentSeeder and change comment rating from decimal to integer

+++ database/migrations/2021_01_20_054857_create_comments_table.php
@@ -16,7 +16,7 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
             $table->text('comment');
-            $table->decimal('rating');
+            $table->integer('rating');
             $table->timestamps();
             $table->foreignId('user_id')->constrained();
             $table->foreignId('company_id')->constrained();
